Prepare for removal of Smartmatch from Perl
This patch silences the 'deprecated::smartmatch' warning on perl 5.38.0,
and if perl version is 5.42.0 or later, smartmatch is not used at all.
This should more gracefully handle the smartmatch deprecation as user
code relying on junction matching that uses smartmatch will still get a
depreaction warning starting in 5.38.0, but older perls will not be
impacted.  On 5.38.0 or later, no deprecation warning will be issued
unless actual calling code relies on smartmatch.