skip minimum perl compat testing
While the Test::MinimumVersion module has an option to skip files, the
Dist::Zilla plugin [Test::MinimumVersion] doesn't.

We include conditional support for smart match. The test file
t/smartmatch.t has instances of smartmatch directly in its code, but
this is protected by a compile time skip_all on perl versions that don't
support it. The Test::MinimumVersion module detects this and reports an
error because of course it can't understand conditional, indirect exit
calls. We would skip just this file, but we can't with the current
Dist::Zilla plugin.