exclude cpanfile from build dist
We use the cpanfile as the source for prereq data. But additional
prereqs will be added by Dist::Zilla, such as develop prereqs for xt
tests. This means that the source cpanfile is misleading if included
directly in the output. Attempting to use `cpanm --installdeps .` will
also use the incomplete cpanfile. We could generate a new cpanfile, but
that data is already in the META files so it serves little purpose. And
if we ever used dynamic prereqs, they wouldn't be reflected in a
generated cpanfile either.

There are various other source files, like the dist.ini, that aren't
needed in the generated tarball. They could potentially also be
confusing, as they aren't usable in the form that they are included. But
since they don't interfere with any tooling, we're leaving them in for
now.